+++ CHANGELOG.rst
@@ -3,9 +3,9 @@ Change log for IQDM-PDF
 
 v0.2.3 (TBD)
 ------------
-- Added optional alternates in JSON templates
+- Added optional ``alternates`` in JSON templates
 - Added optional ``numeric`` flag to make sure value is or is not numerical
-- Ignore parsed values that are equal to their column name
+- Added optional ``ignored`` flag to ignore any returned value in this array
 
 v0.2.2 (2021.01.16)
 -------------------